migration of ballot related methods to new file
